backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
ffmpeg-python==0.2.0
filelock==3.15.4
fsspec==2024.6.1
future==1.0.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.4.0
mpmath==1.3.0
networkx==3.3
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
openai==1.42.0
openai-whisper==20231117
pillow==10.2.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
PyJWT==2.9.0
regex==2024.7.24
requests==2.32.3
rfc3986==1.5.0
setuptools==73.0.1
sniffio==1.3.1
SpeechRecognition==3.10.4
sqlparse==0.5.1
sympy==1.13.2
tiktoken==0.7.0
torch==2.2.2+cpu
torchaudio==2.2.2+cpu
torchvision==0.17.2+cpu
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
